refactor: move all store::redis code into the redis submodule
