+++ crates/interledger-store/src/lib.rs
@@ -4,4 +4,5 @@
 
 pub mod account;
 pub mod crypto;
+#[cfg(feature = "redis")]
 pub mod redis;
